+++ Linked List/Node.py
